% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-stack.R
\name{si_connectance}
\alias{si_connectance}
\title{Connectance of a metaweb or matrix}
\usage{
si_connectance(x)
}
\arguments{
\item{x}{An [si_metaweb()], a matrix, or an [si_field()].}
}
\value{
Numeric proportion of realised links.
}
\description{
Connectance of a metaweb or matrix
}
\examples{
si_connectance(matrix(c(1, 0, 1, 1), 2, 2))
}
