% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-metrics.R
\name{si_evenness}
\alias{si_evenness}
\title{Shannon interaction evenness}
\usage{
si_evenness(m)
}
\arguments{
\item{m}{A weighted matrix.}
}
\value{
Evenness on 0--1.
}
\description{
Shannon interaction evenness
}
\examples{
si_evenness(matrix(c(4, 1, 1, 4), 2, 2))
si_evenness(matrix(1, 4, 4))   # perfectly even
}
