% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-stack.R
\name{si_grid_rast}
\alias{si_grid_rast}
\title{Rebuild a SpatRaster from a grid description}
\usage{
si_grid_rast(g)
}
\arguments{
\item{g}{An [si_grid()].}
}
\value{
An empty single-layer `SpatRaster`.
}
\description{
Rebuild a SpatRaster from a grid description
}
\examples{
library(terra)
si_grid_rast(si_grid(rast(nrows = 5, ncols = 5)))
}
