% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-plot.R
\name{si_palette_change}
\alias{si_palette_change}
\title{A perceptually ordered palette for change maps}
\usage{
si_palette_change(n = 11)
}
\arguments{
\item{n}{Number of colours.}
}
\value{
A character vector of hex colours.
}
\description{
Diverging palette centred on zero, for mapping current-to-future change.
}
\examples{
si_palette_change(5)
}
