% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-plot.R
\name{si_palette_metric}
\alias{si_palette_metric}
\title{A sequential palette for network metrics}
\usage{
si_palette_metric(n = 11)
}
\arguments{
\item{n}{Number of colours.}
}
\value{
A character vector of hex colours.
}
\description{
A sequential palette for network metrics
}
\examples{
si_palette_metric(5)
}
