% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-simulate.R
\name{si_phenology_check}
\alias{si_phenology_check}
\title{Internal consistency check for a derived phenology}
\usage{
si_phenology_check(phenology, metaweb)
}
\arguments{
\item{phenology}{An [si_phenology_from_records()] result.}

\item{metaweb}{The observed interaction matrix (an [si_metaweb()] or matrix)
aligned to the same species.}
}
\value{
A list with the number of testable links, the number of violations
  and the violation rate.
}
\description{
An observed interaction requires that both partners were active at the same
time. This function reports the proportion of observed links whose partners
have no month in common, which should be close to zero if the derived
phenology is sound.
}
\examples{
recs <- data.frame(
  plant  = c("Sp_a", "Sp_a", "Sp_b"),
  animal = c("An_x", "An_y", "An_y"),
  eventDate = c("2014-09/2015-03", "2012-11", "2015-08/2015-11"))
ph <- si_phenology_from_records(recs, "plant", "animal", "eventDate")
mw <- si_metaweb(recs[, 1:2], names_A = colnames(ph$A),
                 names_B = colnames(ph$B))
si_phenology_check(ph, mw)
}
