% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-metrics.R
\name{si_web_metrics}
\alias{si_web_metrics}
\title{All network metrics for one matrix}
\usage{
si_web_metrics(m, what = "all")
}
\arguments{
\item{m}{A weighted or binary matrix.}

\item{what}{Character vector of metric names, or `"all"`.}
}
\value{
A named numeric vector.
}
\description{
All network metrics for one matrix
}
\examples{
si_web_metrics(si_local(si_example_field(), 3))
}
