% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-metrics.R
\name{si_wnodf}
\alias{si_wnodf}
\title{Weighted nestedness (WNODF)}
\usage{
si_wnodf(m)
}
\arguments{
\item{m}{A weighted matrix.}
}
\value{
Numeric WNODF on 0--100.
}
\description{
The quantitative extension of NODF (Almeida-Neto & Ulrich, 2011): paired
overlap counts only those cells in which the less-generalist species has a
strictly smaller weight than the more-generalist one.
}
\examples{
w <- matrix(c(5, 3, 1, 0,  4, 2, 0, 0,  3, 0, 0, 0,  1, 0, 0, 0), 4, 4,
            byrow = TRUE)
si_wnodf(w)
}
\references{
Almeida-Neto, M. & Ulrich, W. (2011) A straightforward computational
approach for measuring nestedness using quantitative matrices.
*Environmental Modelling & Software*, 26, 173--178.
}
